Refactor how datasets are derived at each group level
One of the major changes is that layouts are able to compute "derived"
attributes given a set of options and a dataset. This allows
calculations that involve an expensive initialization step and cheap
incremental steps to not have to undertake the former every time. (For
example, the faceted browser won't have to rebuild a faceted
classification every time a set of options changes)